+++ RELEASE-NOTES.md
@@ -3,6 +3,7 @@
 * #64 jetty-instrumented: handle RequestTimer within the servlet spec
 * #65 Log request execution time outside of jetty-instrumented
 * #66 Support for listener ordering
+* #67 Integrate "business transaction" ids in jetty-instrumented requests
 
 ## 0.24